// src/enums/nutritionsEnum.ts
export enum InputType {
  NUMBER = 'number',
  SELECT = 'select',
  TEXT = 'text',
}

export enum enumNtrCode {
  FAT = 'FAT',
  FASAT = 'FASAT',
  FATRN = 'FATRN',
  CHOLE = 'CHOLE',
  NA = 'NA',
  CHOCDF = 'CHOCDF',
  FIBTG = 'FIBTG',
  SUGAR = 'SUGAR',
  PROCNT = 'PROCNT',
  VITD = 'VITD',
  CA = 'CA',
  FE = 'FE',
  K = 'K',
}